enhancement(wasm transform): Implement some UX improvements for WASM
