+++ lib/vector-wasm/src/lib.rs
@@ -4,5 +4,7 @@ mod registration;
 pub use registration::Registration;
 mod role;
 pub use role::Role;
+mod wasm_module_config;
+pub use wasm_module_config::WasmModuleConfig;
 pub mod hostcall;
 pub mod interop;
