+++ src/wasm/mod.rs
@@ -9,15 +9,13 @@ use crate::{internal_events, Event, Result};
 use lucet_runtime::{DlModule, InstanceHandle, Limits, MmapRegion, Region};
 use lucet_wasi::WasiCtxBuilder;
 use lucetc::{Bindings, Lucetc, LucetcOpts};
-use serde::{Deserialize, Serialize};
 use std::collections::LinkedList;
 use std::{
-    collections::HashMap,
     fmt::Debug,
     fs,
-    path::{Path, PathBuf},
+    path::{Path},
 };
-use vector_wasm::{Registration, Role};
+use vector_wasm::{WasmModuleConfig, Registration, Role};
 mod artifact_cache;
 mod fingerprint;
 pub use artifact_cache::ArtifactCache;
@@ -26,61 +24,6 @@ pub use fingerprint::Fingerprint;
 mod context;
 
 pub mod hostcall; // Pub is required for lucet.
-
-pub mod defaults {
-    pub(super) const ARTIFACT_CACHE: &str = "cache";
-    pub(super) const HEAP_MEMORY_SIZE: usize = 16 * 64 * 1024 * 10; // 10MB
-}
-
-/// The base configuration required for a WasmModule.
-///
-/// If you're designing a module around the WasmModule type, you need to build it with one of these.
-#[derive(Derivative, Clone, Debug, Deserialize, Serialize)]
-#[derivative(Default)]
-pub struct WasmModuleConfig {
-    /// The role which the module will play.
-    #[derivative(Default(value = "Role::Transform"))]
-    pub role: Role,
-    /// The path to the module's `wasm` file.
-    pub path: PathBuf,
-    /// The cache location where an optimized `so` file shall be placed.
-    ///
-    /// This folder also stores a `.fingerprints` file that is formatted as a JSON map, matching file paths
-    /// to fingerprints.
-    #[derivative(Default(value = "defaults::ARTIFACT_CACHE.into()"))]
-    pub artifact_cache: PathBuf,
-    /// The maximum size of the heap the module may grow to.
-    // TODO: The module may also declare it's minimum heap size, and they will be compared before
-    //       the module begins processing.
-    #[derivative(Default(value = "defaults::HEAP_MEMORY_SIZE"))]
-    pub max_heap_memory_size: usize,
-    pub options: HashMap<String, serde_json::Value>,
-}
-
-impl WasmModuleConfig {
-    /// Build a new configuration with the required options set.
-    pub fn new(
-        role: Role,
-        path: impl Into<PathBuf>,
-        artifact_cache: impl Into<PathBuf>,
-        options: HashMap<String, serde_json::Value>,
-    ) -> Self {
-        Self {
-            role,
-            path: path.into(),
-            artifact_cache: artifact_cache.into(),
-            // The rest should be configured via setters below...
-            max_heap_memory_size: defaults::HEAP_MEMORY_SIZE,
-            options,
-        }
-    }
-
-    /// Set the maximum heap size of the transform to the given value. See `defaults::HEAP_MEMORY_SIZE`.
-    pub fn set_max_heap_memory_size(&mut self, max_heap_memory_size: usize) -> &mut Self {
-        self.max_heap_memory_size = max_heap_memory_size;
-        self
-    }
-}
 
 /// Compiles a WASM module located at `input` and writes an optimized shared object to `output`.
 fn compile(
@@ -249,7 +192,10 @@ impl WasmModule {
 #[test]
 fn protobuf() -> Result<()> {
     use serde_json::json;
-    use std::io::{Read, Write};
+    use std::{
+        io::{Read, Write},
+        collections::HashMap,
+    };
     use string_cache::DefaultAtom as Atom;
     crate::test_util::trace_init();
 
@@ -271,6 +217,7 @@ fn protobuf() -> Result<()> {
         "target/wasm32-wasi/release/protobuf.wasm",
         "target/artifacts",
         HashMap::new(),
+        16 * 64 * 1024 * 10, // 10MB
     ))?;
     let out = module.process(event.clone())?;
     module.shutdown()?;
